+++ src/click/termui.py
@@ -675,13 +675,13 @@ def style(
 
     bits = []
 
-    if fg:
+    if fg is not None:
         try:
             bits.append(f"\033[{_interpret_color(fg)}m")
         except KeyError:
             raise TypeError(_("Unknown color {colour!r}").format(colour=fg)) from None
 
-    if bg:
+    if bg is not None:
         try:
             bits.append(f"\033[{_interpret_color(bg, 10)}m")
         except KeyError: